[YS-409] refactor: 파일명 수정
